# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sahaayak",
    "description": "Local-language voice agent for discovering government schemes, scholarships, and jobs, and checking eligibility for them.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sessions/{session_id}/contact-points": {
      "get": {
        "tags": [
          "contact points"
        ],
        "summary": "List Contact Points",
        "operationId": "list_contact_points_api_sessions__session_id__contact_points_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContactPointOut"
                  },
                  "title": "Response List Contact Points Api Sessions  Session Id  Contact Points Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "contact points"
        ],
        "summary": "Create Contact Point",
        "description": "Register a destination and send it a verification challenge.",
        "operationId": "create_contact_point_api_sessions__session_id__contact_points_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactPointCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactPointOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/contact-points/{contact_id}/verify": {
      "post": {
        "tags": [
          "contact points"
        ],
        "summary": "Verify Contact Point",
        "operationId": "verify_contact_point_api_sessions__session_id__contact_points__contact_id__verify_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contact Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactPointOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/contact-points/{contact_id}": {
      "delete": {
        "tags": [
          "contact points"
        ],
        "summary": "Revoke Contact Point",
        "description": "Revoke a contact and stop future delivery to it.\n\nThe row is kept rather than deleted, and an opt-out event is appended. A\ndeleted row would lose the evidence that the caller asked to stop, which is\nexactly the evidence a complaint about unwanted messages turns on.",
        "operationId": "revoke_contact_point_api_sessions__session_id__contact_points__contact_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "contact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contact Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{session_id}/notification-channels": {
      "get": {
        "tags": [
          "contact points"
        ],
        "summary": "List Notification Channels",
        "description": "Which channels this caller can use, and why the others are unavailable.\n\nExists so the reminder form can explain itself. A channel the caller cannot\nuse should say what would make it usable, not present a dead control.",
        "operationId": "list_notification_channels_api_sessions__session_id__notification_channels_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "template_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "benefit_reminder",
              "title": "Template Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelStatusOut"
                  },
                  "title": "Response List Notification Channels Api Sessions  Session Id  Notification Channels Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChannelStatusOut": {
        "properties": {
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "gate": {
            "type": "string",
            "title": "Gate"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "contact_point_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Point Id"
          },
          "display_suffix": {
            "type": "string",
            "title": "Display Suffix",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "channel",
          "available",
          "gate",
          "reason"
        ],
        "title": "ChannelStatusOut"
      },
      "ContactPointCreate": {
        "properties": {
          "channel": {
            "type": "string",
            "enum": [
              "email",
              "sms",
              "whatsapp"
            ],
            "title": "Channel"
          },
          "destination": {
            "type": "string",
            "maxLength": 254,
            "minLength": 3,
            "title": "Destination"
          },
          "locale": {
            "type": "string",
            "maxLength": 8,
            "title": "Locale",
            "default": "en"
          },
          "consent": {
            "type": "boolean",
            "title": "Consent",
            "default": false
          },
          "purpose": {
            "type": "string",
            "enum": [
              "reminders",
              "support"
            ],
            "title": "Purpose",
            "default": "reminders"
          }
        },
        "type": "object",
        "required": [
          "channel",
          "destination"
        ],
        "title": "ContactPointCreate"
      },
      "ContactPointOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "display_suffix": {
            "type": "string",
            "title": "Display Suffix"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "verification_status": {
            "type": "string",
            "title": "Verification Status"
          },
          "consent_status": {
            "type": "string",
            "title": "Consent Status"
          },
          "consent_purpose": {
            "type": "string",
            "title": "Consent Purpose"
          },
          "verified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "channel",
          "display_suffix",
          "locale",
          "verification_status",
          "consent_status",
          "consent_purpose",
          "verified_at",
          "created_at"
        ],
        "title": "ContactPointOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifyRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 12,
            "minLength": 4,
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "VerifyRequest"
      },
...
    }
  }
}